+++ python/test/feature_assembler_test.py
@@ -48,16 +48,16 @@ class FeatureAssemblerTest(unittest.TestCase):
 
         results = self.fassembler.results
 
-        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=4)
+        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=3)
         self.assertAlmostEqual(
-            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
         self.assertAlmostEqual(results[0]["VMAF_feature_adm2_score"], 0.9345149030293786, places=4)
         self.assertAlmostEqual(results[0]["VMAF_feature_ansnr_score"], 23.509571520833333, places=4)
 
         self.assertAlmostEqual(results[1]["VMAF_feature_vif_score"], 1.0, places=4)
         self.assertAlmostEqual(
-            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
         self.assertAlmostEqual(results[1]["VMAF_feature_adm2_score"], 1.0, places=4)
         self.assertAlmostEqual(results[1]["VMAF_feature_ansnr_score"], 31.271439270833337, places=4)
@@ -84,16 +84,16 @@ class FeatureAssemblerTest(unittest.TestCase):
 
         results = self.fassembler.results
 
-        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=4)
+        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=3)
         self.assertAlmostEqual(
-            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
         self.assertAlmostEqual(results[0]["VMAF_feature_adm2_score"], 0.9345149030293786, places=4)
         self.assertAlmostEqual(results[0]["VMAF_feature_ansnr_score"], 23.509571520833333, places=4)
 
         self.assertAlmostEqual(results[1]["VMAF_feature_vif_score"], 1.0, places=4)
         self.assertAlmostEqual(
-            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
         self.assertAlmostEqual(results[1]["VMAF_feature_adm2_score"], 1.0, places=4)
         self.assertAlmostEqual(results[1]["VMAF_feature_ansnr_score"], 31.271439270833337, places=4)
@@ -119,14 +119,14 @@ class FeatureAssemblerTest(unittest.TestCase):
 
         results = self.fassembler.results
 
-        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=4)
+        self.assertAlmostEqual(results[0]["VMAF_feature_vif_score"], 0.44609306249999997, places=3)
         self.assertAlmostEqual(
-            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[0]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
 
         self.assertAlmostEqual(results[1]["VMAF_feature_vif_score"], 1.0, places=4)
         self.assertAlmostEqual(
-            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=4
+            results[1]["VMAF_feature_motion_score"], 4.0498253541666669, places=2
         )
 
         with self.assertRaises(KeyError):